LogUserControl: Add ReadLogSafe(); thanks, Sven
The difference to ReadLog() is that exceptions are caught just like the
button click handler did before. In fact, its code was moved, and it now
calls the new method to do the exception wrapping.

With this, we can make our MainForm constructor crash-safe again,
by simply using the new method.