Add missing BOMs (Byte Order Marks) to the .csproj files
The .sln solution file and all the .cs C# source files have UTF-8 BOMs
already, so adjusting for symmetry. It is not easy to get there with
VIM, as it will add a final '\n' newline to the file in the mode that
supports BOMs, and will not support BOMs in the mode that supports
leaving out the final newline (:set binary). But Visual Studio wants /
MonoDevelop generates .csproj files without final line ending chars,
but VS wants/generates UTF-8 with a BOM. So I now had to use
"git add -i" (interactive) to check in to the index partially, adding
the BOMs and leaving the change of the last line out. ...

(The BOM would normally have been added by the recent Windows .csproj
editing commit, but I had to rebase and manually reedit the two commits
as one introduced CRLFs for the changed lines only (which means the
result is broken), while the other rewrote the whole file to CRLF (which
made an unnecessarily large diff). Saving the latter with VIM after
doing a ":set fileformat=unix" for CRLF -> LF reconvert, ":set binary"
for last line no-line-ending preservation, made a much more compact diff
but didn't insert the BOM anymore (although Visual Studio would have
done so, even when somehow checking in with CRLF converted back to LF,
as I understand it). This commit now makes up for that.)